services/iris-user-state/endpoint/requirements.txt
flask
numpy
scipy
scikit-learn
redis
grpcio
